Hot fix

+++ 0x02-functions_nested_loops/main.h
@@ -1,2 +1,3 @@
-int putchar(char);
+#include <stdio.h>
+int (int)putchar(char);
 void print_alphabet(void);
